Support for custom log file in github-notifier
Currently github-notifier always logs to the working directory.
However, for security reasons it's very common to have cron run
the script as an unprivileged user. In that case, if that user is
not the owner of the directory where the script lies (good
practice if we want to prevent the unprivileged user from
replacing the script itself), it won't be able to write to that
logfile.

With this commit, github-notifier gets a new [-l|--log] command
line option to allow users to use a custom logfile.